Regularize spacing in `texvc` and `texvccheck`
Tweak the position of spaces between function invocations and arguments
so that the expansion of mediawiki-specific functions (`\Rho`, etc) is
consistent with how spacing is added to normal function invocations.

This helps ensure that `texvc`/`texvccheck` is idempotent -- that is,
that it can parse what it emits, and render it without change.

Change-Id: I05fc8fa0201debf611ed4900d3cc953a6a8188ac