Remove unused empty production for `litsq_zq` in texvc
It is impossible to take the empty alternative for this production,
since any input which matches `litsq_aq` would instead match `lit_aq`
if the empty alternative is used.  In addition, there is no action
clause for the empty alternative, so it wouldn't typecheck if the
empty alternative was possible.

This has no effect on the generated parser, but clarifies the grammar.
This commit matches a25078ff9d62ba75ff9ba78eb149a03586154556 to
texvccheck.

Change-Id: I52eddd7321bd8450b47a5d57f2d704f0dd669f16