Update patch set 2
Patch Set 2:

I'm not against this change, but my goal is to improve the Math rendering for the visitors and provide MathML as output to them or if their browsers do not support HTML5 vector graphics.
During the new developments some stress was produced for the opration team.
I don't want to bother them with additonal stress from changes that are valid and correct from a theoretical or philosophical perspective, but have no impact on the rendering for the Wikipedia visitors.
So I don't want to be responsible for performance problems caused by changes of the in my opinion deprecated texvc module.
That said, if I would merge this change and even more the follow up changes that would be against my interests.
The risk I see is that more problems with the PNG rendering will lower the probability that ops support the movement to better MathML + SVG rendering.

Patch-set: 2