Improve PNG fallback
The PNG fallback for old IE versions worked
only if the PNG image was already stored in the
database.
This change enables the Specialpage that displays
the PNG images to create new PNG images and to
show them directly.

To test this commit you need an old IE version
(IE<9). If you don't have access you can use
an online service such as http://netrenderer.com.

Change-Id: I74163da94e2d7a2f07d78bfe0f5ce134564ba97f