Update patch set 2
Patch Set 2:

I think if there is interest in continued PNG support the way to go is to seperate texvccheck from texvc.
If the texvcjs version is good enough it would make sense to use it for mathoid rather than texvccheck which is used for mathoid rendering at the moment.

Patch-set: 2
Label: Verified=0