Update patch set 32
Patch Set 32:

It works on most on the pages that we tested.
But currently MathJax crashes on the CoverageTest page for me
...
unfortunately this page does not render when I open the original
  https://www.mediawiki.org/wiki/Extension:Math/CoverageTest
page in my browser FF24 either. So I need to fix that problem first.

Patch-set: 32
Label: Verified=0