Remove the "best for most browsers" comment for the MathJax mode.
This is a subjective comment that was not true for everybody (trade-off between speed & quality). Moreover, this is now an orthogonal option to enable MathJax client-side rather than a separate rendering mode per se, so it is confusing to compare it against the other modes. The "recommended" mode in the future is likely to become MathML + SVG fallback (which instead uses MathJax server-side) ; but it's probably best to keep neutral regarding the user's preferences.

Change-Id: I5c6ff7c95cded4f06f6bfdef43dbe190e42a2968