Update patch set 68
Patch Set 68:

So I read the MathJax doc & code, and what Amber initially did was correct. It's just that MathJax.Ajax.config.path is only defined in MathJax.js version 2.4 while we are still at version 2.3. So either we wait to upgrade MathJax or we hardcode the path in each js files for now.

Patch-set: 68