Update patch set 5
Patch Set 5:

When you use constants you have to include the extension before declaring the config so that the constant is loaded (or otherwise just hardcode the number the constant equals). For example, initialisesettings.php has this problem. filebackend.php does to and you can see the hardcoded multiwrite constants. It's probably better to just avoid that mistake beforehand in this case. 

You might plan on just dumping the settngs after the Math extension include in CommonsSettings.php which will work, but it still seems like bad practice. Not a blocker I guess though.

Patch-set: 5