Update patch set 5
Patch Set 5:

Thanks for the explanation. I still don't understand how those settings can be set before they are declared. i.e. why are they not overridden once the extension file i.e. Math is loaded. Or does that problem appear only if the extension is disabled.  Do you have an example for a "good practice"?
Do you think it would be possible to use string constants that are compatible to the current boolean value true or false.
I"m really willing to improve that...i.e. I now understand that there will be a problem if the Math extension is disabled for whatever reason. But I don't see a good solution to the problem.

Patch-set: 5