Update patch set 5
Patch Set 5:

I did some research about that.
As far as I understand the configuration should be handled as described here
https://www.mediawiki.org/wiki/Manual:Configuration_for_developers
in the future.
Unfortunately, currently config seems to be a list of key value pairs as setting only. It would be nice to introduce some sort of domain for a configuration variable. That would ensure that extension developers can trust that the config value they are asking for is within a specific domain.
As soon as I understand the handling of config, I'll implement that for the math extension. But I think this would be a follow up commit.

Patch-set: 5
Label: Verified=0