Update patch set 7
Patch Set 7:

OK, so you just mean a temporary fix. But why isn't PNG generated in production and why can't we fix that?

Currently, the MathML/SVG/PNG is chosen in the CSS and I would have expected that "display: none" on <img> prevents the server call for images. So actually, there is another problem here: for browsers with MathML support, we should not load SVG either!

Patch-set: 7