Update patch set 10
Patch Set 10: Code-Review-1

The code looks good to me.  In a number of places you say "new equations".  It should be useful for documentation to better explain what you mean.  I believe the full answer is, "every equation which is rendered by the math extension has its rendering cached in a database.  If the database does not yet contain the given equation, then it is passed through texvccheck."  Assuming this explanation is correct, some version of this should probably be added to the documentation for MW_MATH_CHECK_NEW.  (And how is the database gc'ed?)

In addition, there are potential security issues here, since if you are changing the value from MW_MATH_CHECK_NEVER to MW_MATH_CHECK_NEW there could be arbitrary bad stuff already lurking in the database.  There should probably be some documented way to clear the database before changing the value to MW_MATH_CHECK_NEW.

And are we *sure* that there is no other way an equation could make it into the database unchecked?  I'm not familiar enough with the codebase to be certain enough to +2 this patch.

I'm -1'ing it for the documentation issues, since I believe that security-sensitive settings ought to be very carefully documented.

Patch-set: 10
Reviewer: C. Scott Ananian <664@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1