Update patch set 7
Patch Set 7: Code-Review-1

What happened is that the MathJax mode used to load the TeX source and display its HTML-CSS rendering, but it was very slow. Peter Krautzberger suggested to initially display the PNG images as a "preview" and then progressively replace them with the MathJax's output, and he lobbied to make this the default mode. As a compromised, I suggested to make it an orthogonal option:

1) PNG
2) LaTeX source
3) MathML
4) PNG source as a preview, followed by MathJax (Peter's request)
5) LaTeX source as a preview, followed by MathJax (same as MathJax's original behavior)
6) MathML as a preview, followed by MathJax (same as MathJax's MathML preview in http://docs.mathjax.org/en/latest/options/mml2jax.html). In theory this could speed up the conversion compared to 4) and 5) since LaTeX-to-MathML is done server-side, but in practice MathJax has some issues that makes this slower.

I personally like these options and find this pretty clear, as long as you don't think "MathML + SVG" but "MathML with SVG fallback". This highlights that mode 3) is the standard way with MathML and that SVG is only here as a fallback because of browsers' limitations. This also presents MathJax as an orthogonal technology to render mathematics and enhance the user interface and not a competitor to the MathML standard, as would wrongly make think a parallel option. This also solves the divergence between Peter & Moritz, since we can now agree that mode 3) is the one to choose as the default in the future while still allowing MathJax's enhancement for people who want that.

BTW, I also don't really appreciate that we didn't get much help or feedback while developing the MathML mode and only get attention and changes to long-discussed design decision suddenly because it's about to go in production... and I just realized this morning that mathoid was finally updated to MathJax 2.4 and that this broke the TeX annotation again :-(

Patch-set: 7