Update patch set 7
Patch Set 7:

Generally, the idea that you need to select a rendering mode to account for browser limitations is flawed, as 

a) only logged-in users even have the chance to do so, with only a small fraction actually bothering, and
b) users will use different browsers, even in parallel.

Re-rendering pages based on user preferences is also pretty slow. So overall I'd like us to move to a single HTML representation. This means that eventually any remaining options here will become purely client-side options. I see the separate PNG mode go away completely (it'd automatically be used as a fall-back only). We should also make the rendering on text browsers acceptable without a separate mode. That would basically leave MathML / SVG / PNG as the only mode in the longer term. We can enable MathJax in addition to this, but I don't really see how that makes sense. Zooming can be implemented more cheaply without it. Are there any other reasons to offer it?

In the short term, lets avoid introducing all kinds of weird mode combinations, which are expensive for users (confusing) and developers (need to test all these modes, investigate error reports). I'd even be happy with axing MathJax completely unless there is a good reason to keep it as a separate mode.

Patch-set: 7