Update patch set 7
Patch Set 7:

I agree with you about the long-term goal: keeping only one server-side output with some small client-side ajustments. Currently, maintaining the client-side MathJax mode is painful, it has many big files and is heavy to execute so the idea was rather to move to an independent mode / module for MathJax and defer the maintenance to the MathJax community. Now the question is what is the best strategy in the short term to achieve that goal:

a) Keeping MathJax as a parallel mode, which would mean putting it at the same level as the other server-side modes and thus confusing users about its nature. Also, note again that Peter's request about a mode for MathJax with PNG preview will be lost.
 
b) Keeping MathJax as an orthogonal mode to really show it is a client-side enhancement ; but unfortunately adding many other combinations (which I admit is also confusing to users)...

c) Axing MathJax completely (or rather put it as a separate extension), which will avoid us significant maintenance work but is likely to make the MathJax community and sponsors unhappy.

If I must play MathJax's advocacy, here are the main enhancements it brings:

* Better rendering than the average MathML support in browsers ; even MathJax's HTML-CSS output is also better than its SVG output. We could hope that in the future MathML support in browsers will improve or even that solutions like KaTeX could give HTML-CSS output server-side, but at that point it's not clear.

* Improve accessibily for tools like ChromeVox, TextHelp and JAWS which rely on MathJax's API (if I understood Peter's recent mail correctly). Other tools are currently developed to work directly on MathML like Design Science library for IE+MathPlayer / Firefox+NVDA, Safari+VoiceOver and Orca, so MathJax is not strictly needed but again we might not have full support on all platforms without MathJax at the moment.

* UI enhancements like zooming, font selection, copy and paste. As you said for those who want that, it's easy to implement small scripts or even install browser add-ons like https://addons.mozilla.org/en-US/firefox/addon/mathml-zoom/, https://addons.mozilla.org/en-US/firefox/addon/mathml-copy/, https://addons.mozilla.org/en-US/firefox/addon/mathml-font-settings/. Font selection is not possible for the SVG fallback (unless we generate and cache different output for several math fonts) but it's easy to write a CSS user stylesheet to select the font-family for the MathML and it integrates even better in the MediaWiki way than MathJax's selection menu.

So to summarize as I see the situation, we could avoid MathJax client-side in the long-term but in the short-term it might still be useful to some people.

Patch-set: 7