Update patch set 7
Patch Set 7:

Sounds good to me, especially if we can't make text browsers work well without a separate mode. I'd think that we can do so (alt texts are typically displayed), but we can do so in a second step.

If we do the first change without a change in the labels, then that would help the speed at which we can deploy this. Otherwise we'll have to wait for localizations first.

Patch-set: 7
Reviewer: GWicke <164@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0