Update patch set 1
Patch Set 1: Code-Review+2

To fix the style information for the SVG image was quite complicated. I think it would be better not to touch that again. Especially since displaystyle equations are a new feature that is not in production now.

Patch-set: 1
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2