Update patch set 1
Patch Set 1:

Yes, to get decent MathML rendering, you need a math font (not only for math Unicode symbols but also for stretching fences, drawing large operators etc). The Latin Modern WOFF font should do that, but it is quite big so I'd prefer not to add it for mobile to start with.

(another issue is https://bugzilla.mozilla.org/show_bug.cgi?id=1002526, but it depends on whether the mobile use font inflation)

Patch-set: 1