Update patch set 1
Patch Set 1:

@Physikerwelt: I checked whether the font shows up in the network panel in FF. I guess the reason why I didn't see the load is that I have the fonts already installed locally (thanks @Frédéric).

I just checked the size of the font in my math extension checkout.. it's a whopping 473K. I fear that most FF desktop users will probably end up loading this. That might be fine for the current opt-in mode, but I think that would be excessive for MathML by default.

Patch-set: 1