Move debug integration test to special-page
Since all other LaTexML integration tests will be performed
from the math status special page, it seems reasonable to
run the debug tests from this specialpage as well.

Localisation updates from https://translatewiki.net.

LaTeXML: prevent automatic rerendering of SVG

MathMathML::renderingRequired called the getSvg
function in MathLaTeXML that caused an automatic
rerendering of the SVG image if it was not cached.
That this rendering is triggered from the function
renderingRequired is contra intutive and should be
avoided.

Change-Id: Ieb4776469841fd428f885917d1b98c2b4e823b4b