Update patch set 3
Patch Set 3: Code-Review-1

The problem is that the variable is becoming undefined. Wrapping it in an array will change it from "null" to an array containing "null". That is not an improvement. 

For people experiencing this bug (you), this will change the situation from: Code is not working, variable is undefined. To: Code is not working, value is not found in the array.

For people not experiencing this bug, this change will break Math and change the situation from: The configuration variable is an array, it looks if the value is in the array. To the new situation: The configuration variable is an array, it looks if the value is in the wrapped array. Which will always be false.

php's in_array looks for $search in $values

Example:

 # Before (current code) :
   
 $modes = array( 'png', 'svg' );

 $search = 'foo';
 in_array( $value, $modes );
 // Searching in two values (png and svg): false
 
 $value = 'svg';
 in_array( $value, $modes );
 // Searching in two values: true

 # After

 $modes = array( 'png', 'svg' );
 
 $value = 'foo';
 in_array( $value, array( $modes ) );
 // Searching in ONE value (the $modes array is now one value): false
 
 $value = 'svg';
 in_array( $value, $modes ); // true
 // Searching in one value: false

The code would be broken and not find any of the modes. One might as well remove the code.

Patch-set: 3
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1