Update patch set 4
Patch Set 4:

You don't need to have that in your LocalSettings
The default setting is supposed to be loaded from the Math.php

/**@var array defines the mode allowed on the server */
$wgMathValidModes = array( MW_MATH_PNG, MW_MATH_SOURCE, MW_MATH_MATHML, MW_MATH_MATHJAX )

I still do not understand why it *must* be a problem with the Math extension. I believe that your problem is unrelated to the Math extension.

Patch-set: 4