Update patch set 6
Patch Set 6:

Actually, i'm not skeptical about that. My concern is with rendering after we have generated the snippet, leaving us with a good chance of half the math being rendered and not the other half, because it was cut up by snippeting.

I know Nik and Chad have talked about this before, so I'm sure they have a 'vision' on where this should go longterm, and I'd like to see if this change would simplify, complicate or won't matter to them in reaching those goals.

Patch-set: 6
Label: Verified=0