Update patch set 8
Patch Set 8:

(3 comments)

I'm neutral to +1 on it.  We'll have issues with the snippets being cut off but if we can live with that then so be it.  I'm spinning it up locally to play with it.  Are there any good pages I can import to test it?


One day we could modify our highlighter to be cognisant of <math> tags.  We've already done a lot of work there in the area of highlighting.

Patch-set: 8
Label: Verified=0