Update patch set 8
Patch Set 8:

It looks like Cirrus breaks this feature by removing all the <math> tags from the text.  I imagine this is a side effect of HTML clearing.  I loaded the Fourier Transform page locally with math enabled and hit it with ?action=cirrusdump and saw that the text field is missing all the math.  I imagine we could fix Cirrus to not be silly about this but I'm not 100% clear on what we'd do.  Just not strip the math tags and let the hook strip them on render?  That'd work but might cause trouble with snippets.  I know snippets aren't perfect but this'd probably make them much worse around math tags.  Still, it seems like this is a noble thing to do.

Patch-set: 8