Update patch set 8
Patch Set 8:

For the record I haven't run it without Cirrus but I assume it works there well.  I think the right thing to do is make this a feature flag-able thing and let people that want it turn it on but have a warning that it doesn't currently work with Cirrus.  We can remove the warning and enable it on the cluster if we can fix cirrus.

Patch-set: 8