Update patch set 8
Patch Set 8: Code-Review+1

(1 comment)

OK.  I'm +1 after some trying.  I wonder if we're going to see a little bump in load during the render but that'll probably be fine.

Otherwise its cool.

One thint to be super clear about is that it reverts to spitting out <math> tags if we can't find a start and end math tag.

It'd be super cool if you could make the matched sections bold in the math rendering.

Patch-set: 8
Label: Code-Review=+1