Update patch set 1
Patch Set 1:

There is a backport in the works to bring a compatibility the PSR-3 logger to 1.23.10. When that lands it would probably also be added the 1.24.x version as well. The 1.19.x series will hit end of life next month (May 2015) so a backport there may not be worth the effort.

Patch-set: 1
Label: Code-Review=0