Update patch set 13
Patch Set 13:

I added the random number.
If there is a way to get a pointer to the position in the wikitext source that is supposed to be presented I would love to use that.
This would have pos and cons tough.
+ No excaping and un escaping
+ Render math tags that are cut in the middle now
- Redo highlighting

If someone is aware of another hook that I could use instead, I'd be happy to do that.

Patch-set: 13
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0