Update patch set 5
Patch Set 5:

Yes, and $wgMathValidModes is set to 
  array( MW_MATH_PNG, MW_MATH_SOURCE, MW_MATH_MATHML, MW_MATH_MATHJAX )

I see no place anywhere in Math's master branch (as of 6b62c5d) where it's set to a value other than that (except in unit testing, but that then sets it to another array).

Clearly you've got some path or config that's causing it to be set.

Patch-set: 5