Update patch set 5
Patch Set 5: Code-Review-2

I think in general it would be great to check the config that is set by the user before accessing it. For instance in a vagrant or a shared host without shell access it can be hard to find where a certain wrong configuration was set.
However, I see this as general improvement for future versions of MediaWiki after all config is manged using new json based configuration method. I don't like the idea to use self correcting for invalid settings, and at least there should be a warning that the config is messed up.

Patch-set: 5