Update patch set 5
Patch Set 5:

You've written if ( true ) to unconditionally execute this block.

It fixed your problem because this block had to be executed on your install, and you accidentally written a queer construct doing exactly that.

Let's see more in detail how it works.

This code creates a non-null, non-zero, not-false, not-empty value, ie an array containing this:

[ 'MW_MATH_LATEXML', $wgMathValidModes ]

We could rewrite it like that:

$someArray = [];
$someArray[] = 'MW_MATH_LATEXML';
$someArray|] = $wgMathValidModes;

if ( $someArray ) {
 //...
}

As $someArray is an array with at least one item, it's evaluated as true.

Patch-set: 5