Update patch set 2
Patch Set 2:

I essentially agree with what you say on https://phabricator.wikimedia.org/T99369. However, it seems to me that some people still like the MathJax client-side mode. Did you discuss that with these people? and with the MathJax team?

(an alternative that was discussed at some point was to create a separate extension for MathJax client-side and to let interested people take care of that, although that would need some work to extract it from the Math extension...)

Patch-set: 2