Add composer.json
This file will be used for php code sniffer.

( i will do a follow up to this patch with php code sniffer )

Add js test to grunt file. Ther is already jshint test run but we should
also do it in grunt file since that is the new way to do it and jshint is
the old way.

There is currently no js files but when someone adds some this test will
be run. It makes it easer if we add the test now and someone uploads a js
file. Someone may think it passed the test when it didnt and didnt even
run.

Required-by: https://gerrit.wikimedia.org/r/#/c/228285/

Change-Id: Icb4b53973b19b6e5e799cca20ecdb63853bf5cb3