Update patch set 4
Patch Set 4:

the remaining question is if you want to name the new branch 3.0.0 or just 3.0 at least it should be consisitent
By the way... I did not maintain the version numbers because I think the automatic wmf versions are sufficent. Moreover I had the feeling that different versions would confuse users.

Patch-set: 4