WIP: Restbase Class
DO NOT MERGE... This change is just for discussion
In a first step this change should deomstrate how the
new restbase interface can be used from within wmf and
from private wikis. In a second step this will replace
the current MathML rendering mode.
The new rendering mode 9 (MW_MATH_RESTBASE) is not supposed
to be merged to any long term maintained repository

Change-Id: I764a183bae5b288ec7cb2204ff2af1c339054a95