Update patch set 8
Patch Set 8:

(2 comments)

Cool! Now we have something :) Next step - define the code path for when RB's VRS config is not available. I think the only viable choice here is to fall-back to mathoid, since it is safe to assume that if RB's VRS config hasn't been loaded, the MW instance does not have it installed.

> I'll overwrite the mathml rendering mode with this one, to avoid confusion.

What do you mean by that?

A couple of comments in-lined.

Patch-set: 8
Label: Verified=0