Update patch set 2
Patch Set 2: Code-Review-2

I think it's not a good idea to create tables just because it's easier to program.
MathRenderer::getValidModes() should always return an array. If it does not do so this bug has to be fixed on that end.

Patch-set: 2
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2