Update patch set 1
Patch Set 1:

It seems to pass phpunit but fail in my wiki not sure why. But if you doint like this or doint want to remove this I can instead opt to introduce a config that is enabled by default. I am not sure what the config could be called.

Patch-set: 1