Fix in_array error
This allows us to run the web updater without it causing the whole sql to
be stoped.

Shows this error at the top

<br />
<b>Warning</b>:  array_map() [<a
href='function.array-map'>function.array-map</a>]: Argument #2 should be
an array in
<b>/home/paladox1/public_html/extensions/Math/MathRenderer.php</b> on line
<b>652</b><br />
<br />
<b>Warning</b>:  array_map() [<a
href='function.array-map'>function.array-map</a>]: Argument #2 should be
an array in
<b>/home/paladox1/public_html/extensions/Math/MathRenderer.php</b> on line
<b>652</b><br />

But allows the sql to be run without error for rest.

Bug: T88946
Change-Id: Ie7ff3caab51c965d457dee17de7285acbdd91245