Update patch set 2
Patch Set 2: Code-Review-1

This changes breaks the fundamental logic of what this code does.

MathRenderer::getValidModes() returns an array. It looks if the string is in that array.

If you wrap it in another array, it will never find it because the item will be one layer deeper.

Please fix your installation so that MathRenderer::getValidModes() isn't broken. It's been explained at T88946 numerous times before.

Patch-set: 2