Update patch set 4
Patch Set 4:

This is just a subset for testing, but we were going to use this as the full list of supported commands: https://en.wikipedia.org/wiki/Help:Displaying_a_formula#Formatting_using_TeX - is this the best list?

The SVGs are embedded into the CSS so that they are all brought over in a single request, which I believe requires them to be on the filesystem.

Patch-set: 4
Label: Verified=0