Update patch set 4
Patch Set 4:

The best list of the supported tex commands is here:
https://github.com/wikimedia/texvcjs/blob/master/lib/texutil.js
For regular rendering the SVG images are included via background images as well
https://github.com/wikimedia/mediawiki-extensions-Math/blob/master/MathMathML.php#L378
and the following lines
Suggestions to improve this are more than welcome.
Currently we point to a specialpage that fetches the svg image from mathoid.
In the future this should be changed and we want to get the svg images from restbase which will be extremly fast, for images that are not rendered for the first time.

Patch-set: 4
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0