Update patch set 6
Patch Set 6:

Thanks for the list - we're working on incorporating more commands and would appreciate feedback about coverage and best practices when the time comes.

To avoid polluting the repository with hundreds of SVG files, we are embedding them as data URIs directly in the CSS file. (The browsers that VisualEditor supports support this.)

Patch-set: 6
Reviewer: Tchanders <2849@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0