Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

One minor comment.

I wonder if using VE's config is wise, since that creates an implicit config dependence of the Math extension on VE, in the sense that if VE changes the semantics of the global variable, it effectively becomes impossible for Math to stay compatible with it between versions.

Patch-set: 6
Label: Code-Review=-1