Update patch set 7
Patch Set 7:

@bene you mean, Wikibase should consistently depend on many other extensions? Instead of the other way around?

The clean approach would be to have a separate extension for the glue code. That also makes more sense with regards to versioning. But it's quite a bit of overhead.

In this case, I think we should go for what is most convenient. One problem with having this in the Math extension is that not many people from the Wikidata team have +2 rights here. Otherwise, I don't think it matters much.

Patch-set: 7