Update patch set 22
Patch Set 22: Code-Review-1

(10 comments)

I was wondering how this could ever succeed. And indeed, patch set 22 simply deletes the failing tests! What!?

This patch introduces a lot of dependencies that are not mentioned in composer.json. That's the reason why Jenkins can not run the tests.

This new code makes Math depend on Wikibase. This should and most probably can not be done. Users of the Math extension should not be forced to install Wikibase when they neither care about nor need Wikibase. This new code must be in a separate extension.

I would -2 this patch for the reasons explained but I do not have the right to do so.

Patch-set: 22