Update patch set 22
Patch Set 22:

Hi, thank you for your code review.
"This patch introduces a lot of dependencies that are not mentioned in composer.json. That's the reason why Jenkins can not run the tests."
Can you be explicit about that.
There is also no depency to VE, even thought there are VisualEditor parts in the Math extension that are only relevant if the VE extension is activated.
I'd very much like to treet those two dependencies in the same way.
See also T122656

Patch-set: 22