Update patch set 22
Patch Set 22: Code-Review-2

The current state of patch set 22 shows a possible way, but it's bad: It's impossible to have any tests for code that relies on undeclared dependencies. Apart from that, it will work as intended. For a while, at least. It can and will break any time something in Wikibase or one of the other libraries changes, and you will not notice until something in production (on Wikipedia, in the worst case) breaks.

See https://github.com/wikimedia/mediawiki-extensions-Wikibase/blob/master/composer.json for an example how to declare dependencies.

I would like to discuss this in the Wikibase team. I'm setting a temporary -2 until this happened. Currently the guys are in San Francisco.

Patch-set: 22