Update patch set 26
Patch Set 26:

(2 comments)

Thanks for restoring the tests! I think what Daniel said about "soft dependencies" is ok and the way to go.

One think you could do is to add all the new dependencies introduced in this patch ("data-values/validators": "~0.1.0" and so on) to the "require-dev" section in your composer.json. This should make it easier to run the tests in an environment that does not have Wikibase pre-installed. But this is optional.

Patch-set: 26
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0