Update patch set 2
Patch Set 2:

Since when can qqq changes be made on translatewiki? And what will happen then? Years ago I tried to edit en messages on translatewiki and got reverted with very angry, non-helpful comments stating that I'm not allowed to do this (even if the wiki allows it for a reason I still do not understand). How is this different with the qqq messages? How are the qqq.json files in the repository updated? Or will they be out of sync forever? Sorry, doesn't make any sense to me at the moment.

Patch-set: 2