Update patch set 2
Patch Set 2:

The error is caused by bad untested code introduced to core a few hours ago. I added you both to I17ed5b6.

About the revert: I'm trying hard to fix stuff myself when I can, but my time is limited. I wanted to -1 this patch when I saw the error. This would have been the easiest, least frustrating, most time-saving way. With the fast, erroneous merge I had three options: Wait for somebody else to fix it, fix it myself, or create a revert so it can not be forgotten. I did not expected you to *ALSO* merge this immediately. Again, this is not helpful. We could have updated this revert patch with a 2nd patch set. Now, this is *AGAIN* not possible. :-(

Again, please let people review patches. A good rule of thumb is to wait 1 hour.

Patch-set: 2