Update patch set 2
Patch Set 2:

If MathML is so much better, why don't we have a MathML property type? Again, what's the benefit of having this conversion here, in the RDF export. What if a user of the RDF wants the original input? What if a user of the RDF output finds a mistake in the value and wants to make an edit on Wikidata? How and where will the MathML be converted back?

Patch-set: 2