Make short mode for the math dialog
When the math dialog's layoutMenu height is below 450 px,
make the symbol menu and the preview element shorter, so
the input is still in view. (This works until about
280px, below which the user will need to scroll down to
use the input.

NB Nested setTimeouts were needed in order to check the
space available for the input after the symbols menu
transition has finished. This means that resizing the
dialog multiple times is slow. It is probably rare for
a user to change the height of the dialog while using
it, however.

Bug: T121150
Change-Id: I9e41cce80c3f770b1e3d0e7b133470045a1afe67