Update patch set 1
Patch Set 1:

The extension you mention won't work for MediaWiki as I have to regenerate the binary. Unfortunately my dev certificate expired and renewal is no longer free, so I don't think I'll do it. Safari has a feature to set some custom CSS so that should not be a problem.

Patch-set: 1