Update patch set 2
Patch Set 2:

The profiling data is there just for my convenience;-)
The first link does not point to the intended version of the page.
I intended to create a page with the following content
<math>\textstyle{}</math>
However, since the math extension is currently throwing an exception, the bad input can not be stored.

Patch-set: 2
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0