Update patch set 2
Patch Set 2:

(2 comments)

The other two are fine but to make the rule less generic,
As there are no parent selectors in CSS, I'll probably have to wrap all the math equations in another div or give the "dl"s containing Math, a class. 
But for that, I haven't been able to figure out what php file is generating the HTML output for the Math formulae.
How exactly do I find that to wrap the math contents?
Thanks :)

Patch-set: 2