Update patch set 6
Patch Set 6:

(1 comment)

I've installed the Math extension locally and tested your change. We need to do a few more tweaks. Please see inline for one of them. We also need to change the MathSource.php and Mathtextvc.php files and add the "mwe-math-element" class to getHtmlOutput() and getMathImageHTML() files respectively. Since we'll be adding this class in 3 different places, I suggest we create a config variable called "MathElementClassName" and add it under the "config" key of extension.json. Then use that config variable in places where you need to add the class.

Patch-set: 6
Label: Verified=0