Do not ignore all .mli files in texvccheck
The pattern "*.mli" was incorrectly added to texvccheck/.gitignore by
4c9dc2f08e03ad23e1e2f0cefc52ca6893fbe6a2 with the comment "Compiled
source".  As with d9a7da14f10f6c53bdf1a18febab030d48f0a960, .mli files
define the interfaces for the corresponding .ml module and thus should
not be ignored.

The only exception to this is parser.mli which is generated by the
ocamlyacc from the parser.mly definition, so this change only ignores
those files.

Change-Id: I263df70e58fb14a7228c32d5c985ff7f1b841891