Allow end user to enable or disable extensions w/ ext. dependencies
Because of the way the array_merge strategy works for extension.json,
if a config is set to an array in extension.json, it cannot be
overriden in LocalSettings.php.  This change removes those modes that
have an external dependency so that the wiki administrator can add
only those modes supported on his local installation.

Bug: T150011
Change-Id: I8214b636020b87e22de93519c6c91347ba5b0509