Allow end user to enable or disable extensions w/ ext. dependencies
Because of the way the array_merge strategy works for extension.json,
if a config is set to an array in extension.json, it cannot be
overriden in LocalSettings.php.

This change disables those modes that depend on external services by
default (mathml and latexml).  They can be enabled by using

    $wgMathValidModes = [ 'latexml' => true, 'mathml' => true ];

Bug: T150011
Change-Id: I5622eca17af6d718b6fbb1919e0b970e9a506e67