Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure if I understand the intention... However, why not leaving the defaults empty and set it in the hook if the variable is not set? And why don't we fix the array merge strategy. That way also other extension could benefit from this fix. Moreover, we need to check that mathml is in valid modes for the production cluster. I'm not sure if thats set explitly somwhere or if production relies on the default config.

Patch-set: 2
Label: Code-Review=-1